Fixes for emacs 23.3.
The behavior of eieio's :initform has changed.
We can't use (lambda) as the :initform anymore, and
we can't call (error) to signal an error when the argument is not
given.
Instead, overriding initialize-instance would seem to be the correct
solution for classes which need special behavior at instantiation-time.
This patch removes problematic :initforms
and overrides initialize-instance where necessary.