Always set deleted-chars to the given text
There may be multiple advice functions invoked when deleting a
character interactively.  The current code sets deleted-chars to t in
such a case, which means the deleted character is not spoken.

I don't know what was the purpose of setting deleted-chars to t on a
repeated invocation, but not speaking deleted characters is wrong.
Let's always set deleted-chars to the given text, which should be the
same on contingent multiple invocations.