Fix spelling not working in speech output
The problem with spelling is that SSIP doesn't allow setting it inside blocks.
speechd-out-text always wraps the text and a contingent attached icon by an
outer block, which effectively prohibits spelling (and other parameters not
allowed in SSIP blocks) from being set.

Let's use the outer block only if there is an icon.  This is not a complete fix,
texts with attached icons still aren't spelled, but it's the easiest fix for
now.  A complete fix would be much more complicated.